move codes for order encoding
