+++ cspuz_core/src/encoder/log.rs
@@ -3,9 +3,10 @@ use std::collections::VecDeque;
 use super::direct::{
     encode_linear_eq_direct_from_info, DirectEncoding, LinearInfoForDirectEncoding,
 };
+use super::order::{LinearInfoForOrderEncoding, OrderEncoding};
 use super::{
     encode_linear_eq_mixed_from_info, encode_linear_ge_mixed_from_info, new_var, new_vars_as_lits,
-    ClauseSet, EncoderEnv, LinearInfo, LinearInfoForOrderEncoding, LinearLit, OrderEncoding,
+    ClauseSet, EncoderEnv, LinearInfo, LinearLit,
 };
 use crate::arithmetic::{CheckedInt, CmpOp, Range};
 use crate::norm_csp::{IntVar, IntVarRepresentation, LinearSum};
