+++ cspuz_core/src/encoder/order.rs
@@ -1,0 +1,130 @@
+use super::{EncoderEnv, ExtendedLit};
+use crate::arithmetic::{CheckedInt, Range};
+use crate::norm_csp::LinearSum;
+use crate::sat::Lit;
+
+/// Order encoding of an integer variable with domain of `domain`.
+/// `vars[i]` is the logical variable representing (the value of this int variable) >= `domain[i+1]`.
+pub(super) struct OrderEncoding {
+    pub domain: Vec<CheckedInt>,
+    pub lits: Vec<Lit>,
+}
+
+impl OrderEncoding {
+    pub fn range(&self) -> Range {
+        if self.domain.is_empty() {
+            Range::empty()
+        } else {
+            Range::new(self.domain[0], self.domain[self.domain.len() - 1])
+        }
+    }
+}
+
+/// Helper struct for encoding linear constraints on variables represented in order encoding.
+/// With this struct, all coefficients can be virtually treated as 1.
+pub(super) struct LinearInfoForOrderEncoding<'a> {
+    pub coef: CheckedInt,
+    pub encoding: &'a OrderEncoding,
+}
+
+impl<'a> LinearInfoForOrderEncoding<'a> {
+    pub fn new(coef: CheckedInt, encoding: &'a OrderEncoding) -> LinearInfoForOrderEncoding<'a> {
+        LinearInfoForOrderEncoding { coef, encoding }
+    }
+
+    pub fn domain_size(&self) -> usize {
+        self.encoding.domain.len()
+    }
+
+    /// j-th smallest domain value after normalizing negative coefficients
+    pub fn domain(&self, j: usize) -> CheckedInt {
+        if self.coef > 0 {
+            self.encoding.domain[j] * self.coef
+        } else {
+            self.encoding.domain[self.encoding.domain.len() - 1 - j] * self.coef
+        }
+    }
+
+    #[allow(unused)]
+    fn domain_min(&self) -> CheckedInt {
+        self.domain(0)
+    }
+
+    pub fn domain_max(&self) -> CheckedInt {
+        self.domain(self.domain_size() - 1)
+    }
+
+    /// The literal asserting that (the value) is at least `domain(i, j)`.
+    pub fn at_least(&self, j: usize) -> Lit {
+        assert!(0 < j && j < self.encoding.domain.len());
+        if self.coef > 0 {
+            self.encoding.lits[j - 1]
+        } else {
+            !self.encoding.lits[self.encoding.domain.len() - 1 - j]
+        }
+    }
+
+    /// The literal asserting (x >= val) under the assumption that x is in the domain.
+    pub fn at_least_val(&self, val: CheckedInt) -> ExtendedLit {
+        let dom_size = self.domain_size();
+
+        if val <= self.domain(0) {
+            ExtendedLit::True
+        } else if val > self.domain(dom_size - 1) {
+            ExtendedLit::False
+        } else {
+            // compute the largest j such that val <= domain[j]
+            let mut left = 0;
+            let mut right = dom_size - 1;
+
+            while left < right {
+                let mid = (left + right) / 2;
+                if val <= self.domain(mid) {
+                    right = mid;
+                } else {
+                    left = mid + 1;
+                }
+            }
+
+            ExtendedLit::Lit(self.at_least(left))
+        }
+    }
+}
+
+pub(super) fn encode_linear_ge_order_encoding_native(env: &mut EncoderEnv, sum: &LinearSum) {
+    let mut info = vec![];
+    for (&v, &c) in sum.iter() {
+        assert_ne!(c, 0);
+        info.push(LinearInfoForOrderEncoding::new(
+            c,
+            env.map.int_map[v].as_ref().unwrap().as_order_encoding(),
+        ));
+    }
+
+    let mut lits = vec![];
+    let mut domain = vec![];
+    let mut coefs = vec![];
+    let constant = sum.constant.get();
+
+    for i in 0..info.len() {
+        let mut lits_r = vec![];
+        let mut domain_r = vec![];
+        for j in 0..info[i].domain_size() {
+            if j > 0 {
+                lits_r.push(info[i].at_least(j));
+            }
+            domain_r.push(info[i].domain(j).get());
+        }
+        lits.push(lits_r);
+        domain.push(domain_r);
+        coefs.push(1);
+    }
+
+    env.sat.add_order_encoding_linear(
+        lits,
+        domain,
+        coefs,
+        constant,
+        env.config.order_encoding_linear_mode,
+    );
+}
